+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## [0.1.1] - 25.04.2021
 
 * Updated to FFI 1.0.0.
+
+## [0.1.2] - 06.05.2021
+
+* Calls init in own isolate.
+* Set socket to non-blocking.